pyproject.toml
[project]
name = "fys-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "ronakrpanchal", email = "rhtpanchal76@gmail.com"},
]
dependencies = [
    "langchain-core",
    "langchain-groq",
    "pydantic",
    "fastapi",
    "python-dotenv",
    "pydantic-settings>=2.9.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.scripts]
start = "uvicorn app.main:app --reload"


[tool.pdm]
distribution = false
